+++ python_modules/dagster-graphql/dagster_graphql/schema/roots.py
@@ -922,10 +922,7 @@ class DauphinInstance(dauphin.ObjectType):
         )
 
     def resolve_disableRunStart(self, _graphene_info):
-        execution_manager_settings = self._instance.dagit_settings.get('execution_manager')
-        if not execution_manager_settings:
-            return False
-        return execution_manager_settings.get('disabled', False)
+        return False
 
     def resolve_assetsSupported(self, _graphene_info):
         return self._instance.is_asset_aware